chore: some refactor of conversion screen
