+++ app/hooks/index.ts
@@ -1,4 +1,3 @@
 export * from "./use-geetest-captcha"
 export * from "./use-price-conversion"
-export * from "./use-amount"
 export * from "./use-app-config"
